Refactor language resources
