+++ locales/ko.js
@@ -1,0 +1,28 @@
+module.exports = {
+  system: `당신은 허구의 명언과 그 맥락을 전문적으로 창작하는 AI 명언 작가입니다.
+지정된 tone(분위기)에 맞는 세계관과 어조로 여러 명언과 그 배경 정보를 창작해 주세요.
+각 명언은 다음과 같은 엄격한 형식을 따라야 합니다:
+
+명언 : (따옴표 없는 짧은 문장)
+캐릭터 이름 : (명언을 말한 허구의 인물 이름)
+작품명 : (해당 캐릭터가 등장하는 허구의 작품명)
+연도 : (작품의 시대 설정, tone과 일치하도록)
+---
+
+주의사항:
+- 실존하는 인물이나 작품을 사용하지 마세요.
+- "허구의", "화자" 등의 설명적 어구를 포함하지 마세요.
+- 명언에 따옴표를 사용하지 마세요.
+- 각 명언 뒤에 반드시 "---"로 구분해 주세요.`,
+
+  createBatchPrompt: (tone, count) => `tone: ${tone}에 맞는 분위기로, 위의 출력 형식에 따라 ${count}개의 명언과 캐릭터 정보를 생성해 주세요.
+각 명언 뒤에 반드시 "---"를 넣어 구분해 주세요.
+한국어로 출력해 주세요.`,
+
+  patterns: {
+    quote: /명언\s*:\s*(.*?)(?:\n|$)/m,
+    user: /캐릭터 이름\s*:\s*(.*?)(?:\n|$)/m,
+    source: /작품명\s*:\s*(.*?)(?:\n|$)/m,
+    date: /연도\s*:\s*(.*?)(?:\n|$)/m
+  }
+};
