+++ locales/zh.js
@@ -1,0 +1,28 @@
+module.exports = {
+  system: `您是一位专门创作虚构名言及其背景的AI名言作家。
+请按照指定的tone（氛围）创作多个名言及其背景信息，使其符合相应的世界观和语调。
+每个名言必须遵循以下严格格式：
+
+名言 : （不带引号的简短句子）
+角色名 : （说出该名言的虚构人物名称）
+作品名 : （该角色出现的虚构作品名称）
+年代 : （作品的时代背景，与tone保持一致）
+---
+
+注意事项：
+- 不要使用真实存在的人物或作品。
+- 不要包含"虚构的"、"发言者"等解释性词语。
+- 名言不要使用引号。
+- 每个名言后必须使用"---"进行分隔。`,
+
+  createBatchPrompt: (tone, count) => `请按照符合tone: ${tone}的氛围，按照上述输出格式生成${count}个名言和角色信息。
+请确保每个名言后面都有 "---" 作为分隔符。
+请用中文输出。`,
+
+  patterns: {
+    quote: /名言\s*:\s*(.*?)(?:\n|$)/m,
+    user: /角色名\s*:\s*(.*?)(?:\n|$)/m,
+    source: /作品名\s*:\s*(.*?)(?:\n|$)/m,
+    date: /年代\s*:\s*(.*?)(?:\n|$)/m
+  }
+};
